feat:添加数据

+++ src/main/java/com/deepsoft/haolifa/dao/repository/ProductMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface ProductMapper {
-    long countByExample(ProductExample example);
+    int countByExample(ProductExample example);
 
     int deleteByExample(ProductExample example);
